automation-preflight/requirements.txt
netmiko>=4.3.0
pyyaml>=6.0
pydantic>=2.0.0
packaging>=21.0
